Add valid certificate for *.metrics. hostname
DNS already sends requests for `redis.metrics.${SYSTEM_DNS_ZONE_NAME}`
to the system ALB and then on the right PaaS app. However the ALB
listener only has a certificate for `*.${SYSTEM_DNS_ZONE_NAME}` and so
a TLS error occurs when making a request.

This commit adds a new ACM certificate for
`*.metrics.${SYSTEM_DNS_ZONE_NAME}` and appends it to the certificates
accepted by the listener on the system ALB.