Revert "Allow apps 4GB of memory for staging"
This reverts commit 9613667f2ffd0607c50411464e5bf071b5c7ba5a.
In that commit I reconfigured the `dea_next.staging_memory_limit_mb`
property of `capi-release` in order to allow apps more than 1GB of
memory for staging.

The change appeared to work, but it turns out that it only works the
first time a new app is pushed. Existing apps, even ones created since
the change was deployed, still get staged with only 1GB of memory.

This change seemed safe because in experiments I could use more than
org's quota when staging multiple apps at once, and so it would not
interfere with trial orgs which only have 5GB of memory quota.

After some further digging I don't think it's safe to increase this
value, and least not without understanding more. In particular what
concerns me is that the Cloud Controller codebase does try to count
staging app's memory against the org and space quota [1].

Until more research and discussion can happen this change isn't safe
to leave in place over the weekend, in case it breaks a tenant's
staging. Reverting for the time being.

[1] https://github.com/cloudfoundry/cloud_controller_ng/blob/c74a9bf37443428a1a4d32bed159ae7e0f00e7e0/lib/cloud_controller/backends/quota_validating_staging_memory_calculator.rb#L4-L15