Rework user impact dashboard
We have an existing user impact dashboard that served very well in an
environment where we had a numer of big TV screens on which it could be
displayed. Now that we live in a more remote world, where individual team
members screen real estate is limited, we need a different approach to getting
the same information.

This dashboard is an attempt at creating that. It's being introduced as a
second dashboard initially, so we can see it with real data before replacing
the existing one.