DO NOT MERGE: Add `--recreate` to `bosh deploy` to force recreation of VMs
When adding IP tables rules through a script, the rules are added and not
replaced if there's a change to them. Deploying with the `--recreate` option
will force the VMs to get recreated on every deploy, resulting in a clean
slate.

It'll take longer to perform a deploy, but that's a tradeoff for a better dev
time experience.