Maintain a metric for the number of applications as a Prometheus recording rule
The performance data we present has previously used a very heavy query to count
the number of applications at intervals in the last year, which included lots
of aggregation. To improve the performance, we introduce a recording rule which
counts the number of applications using the same aggregations, but on a smaller
dataset, and saves the result as the metric "cf_app:count".