Pipecleaner now show the content of the offending line for Rubocop offenses
Pipecleaner extracts Ruby scripts from Concourse pipelines and passes them to
Rubocop for validation. Unfortunately, the line numbers that Rubocop reports
don't match those in the file, because it isn't aware of the rest of the
Concourse file.

To make it easier to find failures, pipecleaner now outputs the content of the
offending line.